src/transport/mesh_networky.py
import asyncio
import RNS
from datetime import datetime

class ReticulumMeshNetwork:
    def __init__(self):
        self.reticulum = None
        self.peers = {}
        self.routes = {}
        self.announcements = {}
        
    async def initialize(self):
        #\"\"\"Initialize Reticulum mesh network\"\"\"
        self.reticulum = RNS.Reticulum()
        
        # Set up packet handlers
        self.reticulum.register_packet_handler(
            self.handle_packet
        )
        
        # Start peer discovery
        await self.start_peer_discovery()
        
    async def start_peer_discovery(self):
        #\"\"\"Start peer discovery process\"\"\"
        announcement = {
            'type': 'peer_discovery',
            'timestamp': datetime.now().isoformat()
        }
        
        while True:
            await self.announce_presence(announcement)
            await asyncio.sleep(300)  # Announce every 5 minutes
            
    async def announce_presence(self, announcement):
        #\"\"\"Announce presence to network\"\"\"
        self.announcements[self.reticulum.destination_hash] = {
            'timestamp': datetime.now(),
            'data': announcement
        }
        
        # Implement actual Reticulum announcement
        # This is a placeholder for the actual implementation
        
    async def handle_packet(self, packet):
        #\"\"\"Handle incoming packets\"\"\"
        try:
            if packet.type == 'peer_discovery':
                await self.handle_peer_discovery(packet)
            elif packet.type == 'data':
                await self.handle_data_packet(packet)
            elif packet.type == 'route':
                await self.handle_route_update(packet)
        except Exception as e:
            print(f"Error handling packet: {e}")
            
    async def send_to_peer(self, peer_id, data):
        #\"\"\"Send data to specific peer\"\"\"
        peer = self.peers.get(peer_id)
        if not peer:
            raise Exception("Peer not found")
            
        route = self.find_route(peer_id)
        if not route:
            route = await self.discover_route(peer_id)
            
        # Send data through route
        return await self.send_through_route(route, data)
        
    async def find_route(self, peer_id):
        #\"\"\"Find route to peer\"\"\"
        route = self.routes.get(peer_id)
        if route and self.is_route_valid(route):
            return route
            
        # Route not found or invalid
        return None
        
    def is_route_valid(self, route):
        #\"\"\"Check if route is still valid\"\"\"
        # Check last validation time
        if 'last_validated' not in route:
            return False
            
        age = datetime.now() - route['last_validated']
        return age.total_seconds() < 3600  # Valid for 1 hour
        
    async def discover_route(self, peer_id):
        #\"\"\"Discover route to peer\"\"\"
        # Implement route discovery
        # This is a placeholder for the actual implementation
        return None
        
    async def send_through_route(self, route, data):
        #\"\"\"Send data through specific route\"\"\"
        # Implement actual sending
        # This is a placeholder for the actual implementation
        return {
            'status': 'sent',
            'timestamp': datetime.now().isoformat()
        }