requirements.txt
aiofiles>=0.8.0
cryptography>=37.0.0
requests>=2.28.0
asyncio>=3.4.3
aiohttp>=3.8.0
python-jose>=3.3.0
pynacl>=1.5.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
beautifulsoup4>=4.9.3
markdown2>=2.4.0